host_tools/requirements.txt
fastapi==0.117.1
Jinja2==3.1.6
pyserial==3.5
pytest==8.4.2
uvicorn==0.36.0
